street-extractor/pyproject.toml
[tool.poetry]
package-mode = false
name = "street-extractor"
version = "0.0.1"
description = ""
authors = ["ztarbug <kronprinzfriedrich@gmx.de>"]
license = "AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
osmnx = "^2.0.6"
geopandas = "^1.1.1"
matplotlib = "^3.10.7"
contextily = "^1.6.2"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"